simulate trades and test pairs before investing money.
I created this simulator in the hopes I can test the signals without
actually spending any coins. Obviously, if the signal was good you won’t
earn money same as if it was bad you won’t lose. I’ve set a 1% profit to
test if the signal was worth it and decide to buy or not. You can change
the percentage on line 30 change 0.01 to whatever you want 0.01 equals
1% the code is not “professional” as I did this for myself just thought
I’d share perhaps it might be useful to someone as well.